backend-chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain-core==0.1.0
langchain-groq==0.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
neo4j==5.14.0
pinecone>=5.0.0
numpy==1.24.3
google-generativeai==0.5.0